Read Me
In case anyone else has ran into the issues of control ClassNNs changing in some programs.

Here is a little tool that uses MAA/ACC to find and test the path to controls. 

Just drag the cross-hair to the control you want; then test the sending of text or clicks or copy the generated code. 

- Casper